+++ lib/src/Server/Middleware/Validation/ValidationEngine.dart
@@ -5,15 +5,15 @@ class Validation implements BeforeHook {
   Future before(ReqResContext context) {
       
     var versionValue = common.ValidatorByVersion.defaultVersionValidator;
-    var version = context.request.url.variables.path.firstWhere((v) => v.variable is common.Version, orElse: () => null);
+    var version = context.request.route.variables.path.firstWhere((v) => v.variable is common.Version, orElse: () => null);
     if(version != null)
     {
       versionValue = version.value;
     }
     
-    this._validateVariables(context.request.url.variables.path, context, versionValue);
+    this._validateVariables(context.request.route.variables.path, context, versionValue);
     
-    this._validateVariables(context.request.url.variables.query, context, versionValue);
+    this._validateVariables(context.request.route.variables.query, context, versionValue);
   }
   
   void _validateVariables(ReadOnlyMap<UrlInputData> variables, ReqResContext context, String versionValue) {